% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-sci-repel.R, R/geom-label-sci.R,
%   R/geom-text-sci-repel.R, R/geom-text-sci.R
\docType{data}
\name{GeomLabelScinameRepel}
\alias{GeomLabelScinameRepel}
\alias{GeomLabelSciname}
\alias{GeomTextScinameRepel}
\alias{GeomTextSciname}
\title{GeomLabelRepel}
\description{
GeomLabelRepel

GeomTextRepel
}
\keyword{datasets}
